+++ inbm/dispatcher-agent/tests/unit/aota/test_aota.py
@@ -610,13 +610,21 @@ class TestAOTA(TestCase):
     @patch('dispatcher.aota.application_command.Application.identify_package', return_value=SupportedDriver.XLINK.value)
     @patch('shutil.move')
     @patch('os.listdir', return_value=[])
+    @patch('dispatcher.aota.application_command.CentOsApplication.validate_package_type', return_value=("xPM v3.0 bin i386/x86_64"))
     @patch('dispatcher.aota.aota_command.AotaCommand.create_repository_cache_repo')
     @patch('dispatcher.aota.factory.is_cent_os_and_inside_container', return_value=True, device_reboot="Yes")
-    def test_application_centos_driver_update_raise_pass(self, mock_detect_os, create_repo, listdir, mock_move,
-                                                         support_driver, run):
+    def test_application_centos_driver_update_raise_pass(self, mock_detect_os, create_repo, validate_package, listdir, 
+                                                         mock_move, support_driver, run):
         aota = self._build_aota(cmd='update', app_type='application', uri="http://example.com")
         self.assertIsNone(aota.run())
 
+    @patch('dispatcher.aota.application_command.CentOsApplication.cleanup')
+    @patch('inbm_common_lib.shell_runner.PseudoShellRunner.run', return_value=("Debian binary package", "", 0))
+    @patch('dispatcher.aota.aota_command.AotaCommand.create_repository_cache_repo')
+    @patch('dispatcher.aota.factory.is_cent_os_and_inside_container', return_value=True, device_reboot="Yes")
+    def test_application_centos_driver_update_unsupported_driver(self, mock_detect_os, create_repo, validate_package, cleanup):
+        aota = self._build_aota(cmd='update', app_type='application', uri="http://example.com")
+        self.assertRaises(AotaError, aota.run)
 
 if __name__ == '__main__':
     unittest.main()