+++ inbm/Changelog.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
  - HSD 15010715189 - Telemetry service failed to start when unable to find mender file
  - RTC 509436 - [BUG] SOTA in docker container cannot take snapshot in btrfs
  - HSD 15010407028 - Remove <hash_algorithm> tag before sending provisionNode manifest to vision-agent
+ - RTC 495463 - Fix existing driver get uninstalled if AOTA update failed using .deb file package
 
 ### Added
  - (507873) Added support for Ubuntu 21.10